Change Record shorts to ints.

+++ AkkaDistribution.Common/Records.cs
@@ -2,8 +2,8 @@
 {
     public sealed record Manifest(DateTime Timestamp, HashSet<ManifestFile> Files);
     public sealed record ManifestFile(string Filename, string FileHash);
-    public sealed record FilePieceDelivery(string Filename, string FileHash, string FilePiece, short TotalPieces, short Position);
+    public sealed record FilePieceDelivery(string Filename, string FileHash, string FilePiece, int TotalPieces, int Position);
     public sealed record MissingPieces(string Filename, string FileHash, List<string[]> MissingPositions);
     public sealed record ManifestRequest();
-    public sealed record FilePartMessage(string Filename, string FileHash, short TotalPieces, string FilePart, short Position);
+    public sealed record FilePartMessage(string Filename, string FileHash, int TotalPieces, string FilePart, int Position);
 }
